requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
atproto==0.0.55
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
libipld==3.0.0
pip==24.3.1
pycparser==2.22
pydantic_core==2.23.4
pydantic==2.9.2
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
websockets==13.1
